pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pl-generator"
version = "0.1.0"
description = "Auto-generate P&L Excel from business plans using LLM extraction"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openpyxl>=3.1.0",
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.9.0",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "openai>=1.0.0",
    "streamlit>=1.28.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
simulation = ["xlwings>=0.30.0"]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
plgen = "src.cli.main:main"

[tool.setuptools.packages.find]
include = ["src*"]
